Multi-threading for benchmark tool
This change makes it possible to run a particular algorithm in many parallel
threads. This puts more stress on the interconnect as any time the network used
to be idle can now be used by another thread. I collapsed the bandwidth reports
into a single metric reporting average bandwidth over the entire run.