Take CUDA peer access into account for on-device reduction
The NVLink cube mesh architecture has partial peer access between devices. Two
groups of 4 GPUs have full peer access and every GPU in one group has peer
access to a corresponding GPU in the other group. When we reduce between all 8
GPUs, tree reduction using peer access must be done separately in those groups,
followed by a reduction across any one of the pairs connecting the groups. This
change refactors CudaDeviceReduce to work with this topology.

Additionally, it shuffles device pointers to randomize which GPUs run the
reduction and which communication links are used. The goal of randomization
here is to prevent excessive load on any single GPU or link.