Make CUDA algorithms use stream associated with scratch pointer
These algorithms still defaulted to the stream of the first pointer.
When GPUDirect is used, we'll use a scratch buffer that is closest to
the networking adapter to minimize intra-machine communication.
However, since streams are pinned to devices, we have to use the
stream associated to the device that hosts the scratch buffer.