Specify interface name when creating TCP transport
Summary:
Creating gloo::transport::Device for the TCP transport would default
to using a machine's hostname to figure out what IP address to bind
to. This is not always desirable, for example when the machine's
hostname is not fully qualified, resolves to localhost, or if you want
to use a non-default interface in case of multi-homed machines. This
change allows users to specify the interface name. In a homogeneous
cluster these interface names are typically identical across machines.
This makes it much easier to use -- for example -- the first IPoIB
interface on every machine.

Reviewed By: romain-intel

Differential Revision: D5603962

fbshipit-source-id: fd976d2c6804ea599862b60d53c05dafa302f7ba