Include memory region sge on every receive work request
In all uses of an ibverbs pair, it would first exchange memory
regions for all buffers, before moving on to send/recv on those
buffers. Receives for memory regions and RDMA receives were posted
separately and because they were not interleaved this worked fine.

When these receive operations were interleaved, I saw received memory
regions equal to all NULs. Treating these two types of receives
differently caused a mismatch between posted receives on the
completion queue and the list with memory regions that the code
expected to take writes. And because all inbound writes are one-sided,
we never know whether we have to expect a memory region write or an
RDMA write.

This change merges the approaches so that we only deal with a single
type of receive, that opportunistically includes a scatter/gather list
with the memory address of a memory region. That way, if the remote
side sends a memory region, it can be received, and if it performs an
RDMA write, it will simply be ignored. To map receive work requests to
memory regions, we keep track of the number of receive work requests
posted and index into a fixed array of memory regions. This array can
be fixed because we configure an upper bound to the number of pending
read work requests (kRecvCompletionQueueCapacity).