Ensure an ibverbs buffer has no in-flight sends when destructing
If it is destroyed before all sends have completed, the pair will try
and call handleCompletion on a destructed buffer.

This is treating the symptom and not the cause. It would be better to
implement a deregistration mechanism such that the pair forgets about
the buffer altogether, as well as mechanism for the device to forget
about a pair upon destruction. I'll leave this as a todo.