Remove bundled version of NCCL
NCCL1 is deprecated in favor of NCCL2 and if anyone wants to still use
it they can specify to do so via NCCL_ROOT_DIR.

This change removes the NCCL submodule and makes use of NCCL a compile
time option and constant. If USE_NCCL=ON, then all device side
reductions are performed using NCCL. Otherwise, they are performed with
a tree of CUDA reductions. Calls to cudaDeviceReduce or
cudaDeviceBroadcast are handled based on the USE_NCCL constant.

If Gloo is built as part of Caffe2, it will pick up the NCCL version
bundled with Caffe2 if it is configured to use it.