Set SO_LINGER to avoid TIME_WAIT after close()
Summary:
The test suite contains many tests and creates many connections. We've
reached a point where I see the test suite hanging waiting for ports
to listen on to become available, with many connections hanging around
in the TIME_WAIT state. Setting SO_LINGER with a linger value of 0
means that a subsequent call to close(2) sends a RST instead of going
through the normal FIN flow. This is a bit hacky, but now the tests
can run without hanging or fail with bind(2) errors.
Closes https://github.com/facebookincubator/gloo/pull/96

Reviewed By: plapukhov

Differential Revision: D5982341

Pulled By: pietern

fbshipit-source-id: 787c189477c7b9af457c2941ef7b8ec42cea6b4e