Added GPU direct support for Gloo - Broadcast Operation
Summary:
The helper functions in cuda.h and cuda_prviate.h all work with either CudaDevicePointer or CudaHostPointer. So, most of the implementation remains fundamentally the same.

Since this is a CUDA algorithm, GPU pointers will be provided as input to CudaBroadcastOneToAll. So, when using GPUDirect, we don't need to allocate an additional scratch buffer.

Reviewed By: pietern

Differential Revision: D6049454

fbshipit-source-id: a540afc3344743825bd99f32622aa15c6842b55f