define SPEED_UNKNOWN for compilation with anaconda compilers (#127)
Summary:
When compiling torch with anaconda compilers I ran into this error about SPEED_UNKNOWN undefined:

https://github.com/pytorch/pytorch/issues/4499

I couldn't find the proper header to include -- even if found it is likely not portable.
Given that not too many people outside kernel uses this macro (gloo is one of the few results googling with 'SPEED_KNOWN'), it is
probably a good idea to define it if it is not defined on the system.

Note that the comment on torch set the macro to -1.

The correct value appears to be 0, see e.g. this commit to the kernel:
https://patchwork.kernel.org/patch/9573661/
Closes https://github.com/facebookincubator/gloo/pull/127

Differential Revision: D8459077

Pulled By: pietern

fbshipit-source-id: 35828fb90cecafcd45d8e0625a57f57ef2520ac9