Fix overflow in transport
Summary:
Doing a broadcast with 512M elements failed with writev returning
EINVAL (Invalid argument). It turns out that the offset and length
fields in the 'Op' struct in gloo/transport/tcp/pair.cc were both
declared as a regular int and when used with 512M floats (total size
of 2GB) caused a 32-bit integer overflow. Subsequently, writev saw a
negative length argument, and returned EINVAL.

This change fixes that, as well as an overflow when setting the TCP
send buffer. This was subject to the same problem because setsockopt
takes an integer optval. It doesn't make sense to ask the kernel for a
send buffer of size 2GB, so they are limited to max out at 32MB now.

Reviewed By: manojkris

Differential Revision: D8601859

fbshipit-source-id: 0a1490887335412514d0035980216db8d109e6a8