Support offset/length arguments for send/recv
This allows for reusing an unbound buffer for multiple send or
recv operations against non-empty ranges. This is useful when receiving
into the same target buffer from multiple peers, or when doing
double-buffering against a single region.

For example, you can now issue N consecutive recv operations against the
same unbound buffer and get notified as they complete in sequence, as
opposed to only getting notified when a single recv operation for the
entire buffer completes.