Fix bug concerning remote pending send tally
Summary:
Every immediate recv call triggers a notification to be sent to the
peer that is will be receving from. This is needed because all use of
send and recv is unbuffered; i.e. the sender can only send when it
knows the remote side is ready to receive.

The remote send call checks if there is a remote pending receive, and
if so, immediately sends its buffer. The bug here was that is did NOT
send the accompanying notification that causes the remote pending send
tally on the remote side of the pair to increase. The tally for remote
pending sends is only used to trigger a check for local recv-from-any
calls. If there is one, and the slot, and rank of the sending process
match, the recv-from-any can be fulfilled by this send.

With the tally not being updated correctly, the recv-from-any call
could block forever waiting for a remote send, while one or more
remote sends were in fact pending. Because of the incorrect tally (it
being zero or negative), the check if such a send operation would
match an existing recv-from-any did not execute.

Reviewed By: romain-intel

Differential Revision: D10243958

fbshipit-source-id: a08c0dd48905958cac466737099ec7f45f6e0582