Add new style reduce collective (#140)
Summary:
This implements a ring reduction resulting the result scattered across
processes, followed by a gather operation to the root process.

The halving/doubling implementation is slated to be ported to this
stateless style later and provided as an option through RedudeOptions.

Rebase needed after #137 is merged.
Pull Request resolved: https://github.com/facebookincubator/gloo/pull/140

Reviewed By: manojkris

Differential Revision: D10376422

Pulled By: pietern

fbshipit-source-id: 2f32f693b8437cdff82a6528f18e576966984395