Add new style allreduce collective (#143)
Summary:
This implements a ring reduction resulting the result scattered across
processes, followed by an allgather operation to the root process.

It can take multiple inputs and multiple outputs. If the inputs are not
specified it is assumed to be an in place operation on the outputs.
Pull Request resolved: https://github.com/facebookincubator/gloo/pull/143

Reviewed By: manojkris

Differential Revision: D10376434

Pulled By: pietern

fbshipit-source-id: 17a6ab7e02591a351e551c8582f0095041ddcf05