Move all buffer prep to options class
Summary: Having users poke around in a struct directly is not a great API not to mention a potential disaster when changing things in the future. This change turns all options structs into classes where options are set with setter functions. The options class does all preparation of unbound buffers as needed.

Reviewed By: manojkris

Differential Revision: D10449437

fbshipit-source-id: 7f604319711be3ed177d9606e98f201c2a1bbe70