Timeout and error propagation for unbound buffers
Summary:
Update the transport context with a timeout setter/getter instead of
passing it as argument to its createPair function. This way, the
unbound buffers created from the transport context can reuse that same
timeout as default.

The waitRecv/waitSend functions on an unbound buffer now take an
optional timeout argument. If the operation times out, it signals all
pairs in the transport context that a timeout happened, and throws a
more detailed exception itself. All pairs in the transport context
need to be closed when this happens, as we don't know if unbound
buffers will be around after the timeout bubbles up to the
application.

Reviewed By: manojkris

Differential Revision: D13103991

fbshipit-source-id: 3ca7dd445292ddb59e249ea16ba80b0550d18ac5