Use single object to keep track of pending ops
This used to be tracked in the pair instances, but to support
receive-from-any we need a centralized view into all pending send
operations. This class facilitaties that centralized view for both
pending send and receive operations.

Before this change, instances of the gloo::transport::tcp::Pair class
kept track of remote pending send and recv operations, and notified the
transport context of remote pending send operations. Functionality wise
this was fine, but it didn't include cleanup of unused slots, nor did it
allow for adding this easily. By centralizing the pending operation
tally in a single object, there is only a single object to clean up when
the pending operation tally drops to 0.

This resulted in a memory leak proportional to the number of used slots,
as reported in pytorch/pytorch#16144. Confirmed that the repro listed in
that issue no longer shows an increase in memory. Also added a test in
gloo/test/memory_test.cc that performs the same test.