Use weak reference to unbound buffers in pairs
Summary:
The lifecycle of the unbound buffer is controlled by the user.
It may be constructed as value or unique_ptr and destructed when
going out of scope or when unwinding the stack.

For example:

```
{
  int i = 0;
  auto buf = context->createUnboundBuffer(&i, sizeof(i));
  buf->send(1, 0);
  ...
  throw std::runtime_error("Something happened!");
  ...
  buf->waitSend();
}

// At this point, if `waitSend` was not yet called, the I/O
// thread executing the send operation may still call into it.
// The unbound buffer will have been destructed already,
// so we need to make sure the reference gets invalidated.
```

If upon destruction there are pending operations (e.g. a recv
operation timed out), there will be references to this unbound
buffer in the `transport::Pair` instances that could fulfill the
operation. To avoid these pairs calling into the unbound buffer
after it is destructed, we must ensure that these references are
invalidated before the destructor returns. Also, if destruction of
the unbound buffer races with the recv operation completing after
all, we must block the destructor and wait for the operation to
finish. Otherwise, we risk the device thread writing to memory it's
not supposed to.

We solve this by using a shared_ptr of the "this" pointer of the
unbound buffer. This doesn't magically turn the unbound buffer into a
sharable pointer, but it allows for handing out weak_ptr instances to
refer to it. Then, whenever the unbound buffer is used by another
thread, it converts the weak_ptr into a shared_ptr and uses it for a
very short period of time. The wrapper classes defined in
`gloo/common/memory.h` wait for all shared_ptr instances to be
released before returning from its destructor. This will block
indefinitely if the shared_ptr acquired from the weak_ptr stays alive.

Reviewed By: yyetim

Differential Revision: D13662413

fbshipit-source-id: 447257dab3adbb03f68f5bf8cb2aa2046783c22f