Ensure pairs are destructed before their device (#167)
Summary:
As of 27760a38 the pairs refer to their underlying device with a raw
pointer, which introduced a possible destruction order problem.

If the device is only referred to by the transport context, it would be
destructed before the pairs would be destructed, per ordering of members
in `gloo::transport::Context` and `gloo::transport::tcp::Context`. Then
upon destruction of the pairs, they were no longer able to unregister
their file descriptor from the device, because it was already
destructed.
Pull Request resolved: https://github.com/facebookincubator/gloo/pull/167

Reviewed By: manojkris

Differential Revision: D13924588

Pulled By: pietern

fbshipit-source-id: 67bb92e2070aa9cc4fe2575b1c5cd980d7c531ea