Add bcube allreduce algorithm
Summary:
This algorithm implements a hierarchical reduce/scatter followed by
hierarchical allgather. The number of processes in a communication
group must be identical across communication groups at every step in
the hierarchy. Therefore, the number of processes must factorizable.

If the number of processes is a prime number, the execution of this
algorithm is equivalent to an all-to-all reduce/scatter followed by an
all-to-all allgather. This is suboptimal if there is network locality
to take advantage of.

Reviewed By: mrshenli

Differential Revision: D15325539

fbshipit-source-id: c5c35f0113edf2a2f41b0dff82cbd4b309793799