Move AllreduceOptions to detail namespace
Summary:
To force API users to use the setter functions on this class, all
member fields are protected. This also means that legitimate users
must be explicitly listed as friend class or function. This doesn't
scale as we add alternative algorithm implementations. This commit
moves the member fields of the options class to a struct. Populating
the options struct is done through the setter functions on the top
level object.

Reviewed By: vladbelous

Differential Revision: D15325540

fbshipit-source-id: 2b10bdc944e1ddf2b947ffc981c2c4c90668f3fe