Make collectives work with single process context
Summary:
This is a base case where these collectives are usually a nop. For
code simplicity it is better to not have to differentiate between N=1
and N>1 cases. It's easy to support this case here, so we should. This
commit adds the N=1 case to the test cases and fixes the collective
implementations to deal with it where needed.

Differential Revision: D16205644

fbshipit-source-id: 02c83cfc3d68732923bdf39f979a4c9f72fef38f