Make collectives work with single process context (#189)
Summary:
Pull Request resolved: https://github.com/facebookincubator/gloo/pull/189

This is a base case where these collectives are usually a nop. For
code simplicity it is better to not have to differentiate between N=1
and N>1 cases. It's easy to support this case here, so we should. This
commit adds the N=1 case to the test cases and fixes the collective
implementations to deal with it where needed.

Reviewed By: kuttas

Differential Revision: D16205644

fbshipit-source-id: b767abe68ca173188f70cd91ba125fe6332e7717