Log message if bind(2) fails when looking up address for hostname (#191)
Summary:
Pull Request resolved: https://github.com/facebookincubator/gloo/pull/191

This routine is called when initializing a device. It works by
iterating over the set of addresses returned by getaddrinfo(3) and
trying to bind to each of them. This is done to make sure we're not
binding to an IPv6 address on an IPv4 only host and vice versa.

Binding can fail if the machine is out of ports for that address, but
it would raise an error saying it couldn't resolve an address for the
hostname, which is wrong. Instead, it should raise an error saying
that binding to a particular address failed.

Reviewed By: zhaojuanmao

Differential Revision: D16437786

fbshipit-source-id: f90f01c00e04909f4ea3ba0d64776cc421696211