Create libuv transport
This transport is largely modeled after the existing TCP transport.
There is no support for the old style and only for the new style
algorithms (there are no createSendBuffer/createRecvBuffer functions).

Tests can be run by setting GLOO_TEST_TRANSPORT=uv. By default, none
of the tests against this transport run.

This commit adds the basic functionality but more work is needed to
properly deal with errors and timeouts.

It requires libuv to be available. Either it should already be defined
as a build target, if included by another project's build, or it must
be installed and discoverable through pkg-config.