Remove superfluous asserts
Usage of send/recv (including recv-from-any) with nbytes==0 is
allowed. The asserts for a direct send/recv for the offset and nbytes
argument checked that they are >= 0, which is always true since they
are unsigned integers.

The assert for a recv-from-any did NOT yet allow for an empty message.
This commit adds a test for that case and updates its asserts as well.

ghstack-source-id: f285b4c2b154e8dc188c9131eac3b2223f4781bd
Pull Request resolved: https://github.com/facebookincubator/gloo/pull/206