Remove superfluous asserts (#206)
Summary:
Pull Request resolved: https://github.com/facebookincubator/gloo/pull/206

Usage of send/recv (including recv-from-any) with nbytes==0 is
allowed. The asserts for a direct send/recv for the offset and nbytes
argument checked that they are >= 0, which is always true since they
are unsigned integers.

The assert for a recv-from-any did NOT yet allow for an empty message.
This commit adds a test for that case and updates its asserts as well.

Test Plan: Imported from OSS

Differential Revision: D16801687

Pulled By: pietern

fbshipit-source-id: 8bf02bca9534db4c68e7e8bde2cbf260acb64d68