Make aligned_allocator use posix_memalign (#197)
Summary:
Pull Request resolved: https://github.com/facebookincubator/gloo/pull/197

This change makes the aligned allocator use posix_memalign over
memalign, which is a Linux specific call. This change also moves the
aligned allocator to a separate file per the realization that it is
used only in benchmark and test code.

Test Plan: Imported from OSS

Differential Revision: D16801637

Pulled By: pietern

fbshipit-source-id: 5f57496f1c1fae2c40a1f5610485971ff2bdb3fa