Create libuv transport (#212)
Summary:
Pull Request resolved: https://github.com/facebookincubator/gloo/pull/212

This transport is largely modeled after the existing TCP transport.
There is no support for the old style and only for the new style
algorithms (there are no createSendBuffer/createRecvBuffer functions).

This commit adds the basic functionality but more work is needed to
properly deal with errors and timeouts.

It requires libuv to be available. Either it should already be defined
as a build target, if included by another project's build, or it must
be installed and discoverable through pkg-config.

Test Plan: Imported from OSS

Reviewed By: mrshenli

Differential Revision: D17072672

Pulled By: pietern

fbshipit-source-id: 6594a8a580596e86253b3d8783978c40fafef373