Refactor algorithm tests to take transport parameter (#208)
Summary:
Pull Request resolved: https://github.com/facebookincubator/gloo/pull/208

This is required to run the same algorithm tests against multiple
transports. The test suite is comprised of both class algorithms and
function algorithms. The former require fixed buffers, which are only
supported on the tcp backend (and will be phased out). The latter
require the unbound buffers, which are supported on both the tcp and
uv (not yet merged) transports. This means we can't set the transport
globally and forget about it, since it depends on the test whether or
not it works with the selected transport.

Instead, by parameterizing the tests we allow for multiple transports
to be tested in the same test run. If a transport is not
available (e.g. not compiled) then those tests can be skipped.

Side notes:

* There are a series of small modifications included here as well,
  including more consistent variable naming, and a more consistent
  usage of test parameterization generator functions.
* Because this commit touches many lines anyway, I also ran
  clang-format on all files under `gloo/test` to make the style
  consistent.

Test Plan: Imported from OSS

Reviewed By: mrshenli

Differential Revision: D17072673

Pulled By: pietern

fbshipit-source-id: a408a83404ff522d224e7f25bb6c3e3c62def299