Make libuv detection code more robust
The anaconda package for libuv uses version 1.26. There are no
features introduced since then that we rely on, so we can use that as
minimum version.

The static library in the anaconda package is named `libuv.a`. When
libuv is compiled and installed from source it produces a static
library named `libuv_a.a`. We use CMake's `find_file` to figure out
which is used.