Make libuv detection code more robust (#213)
Summary:
Pull Request resolved: https://github.com/facebookincubator/gloo/pull/213

The anaconda package for libuv uses version 1.26. There are no
features introduced since then that we rely on, so we can use that as
minimum version.

The static library in the anaconda package is named `libuv.a`. When
libuv is compiled and installed from source it produces a static
library named `libuv_a.a`. We use CMake's `find_file` to figure out
which is used.

Test Plan: Imported from OSS

Differential Revision: D17153544

Pulled By: pietern

fbshipit-source-id: b552345fec073ee01a19a356762a07d95c1ec8a0