Update and move remote pending operation accounting
Accounting for remote pending operations was the responsibility of the
transport specific context class. This is essentially identical across
transports, though, so this commit moves the logic out of the two
transport implementations and into the base class.

Second, the original approach could lead to starvation on
recv-from-any calls. Because pending operations were counted keyed off
the remote rank, low ranks would take precedence over high ranks. This
commit changes that such that ranks are serviced in the same order in
which their send notifications are received.