Add abortWaitSend/abortWaitRecv to unbound_buffer (#228)
Summary:
Closes https://github.com/facebookincubator/gloo/issues/227.
Adds 2 methods: `abortWaitSend`, `abortWaitRecv` so that we can interrupt a blocking `waitRecv` or `waitSend`, per pietern 's suggestion. This is useful if a gloo process is waiting for a message in one thread, but another thread decides to shutdown and wants to stop it gracefully.

The methods set a boolean variable and notify a waiting condition variable. The condition variable predicate is changed to check this condition. The implementation is both for the tcp and uv transport methods. Unit tests are also added.
Pull Request resolved: https://github.com/facebookincubator/gloo/pull/228

Test Plan: Added relevant unit tests in `send_recv_test.cc`

Reviewed By: pietern

Differential Revision: D18210908

Pulled By: pietern

fbshipit-source-id: 013206a77edd7122c9ff6a120f3fbf41d5e75a56