Add abortWaitSend/abortWaitRecv to unbound_buffer (#232)
Summary:
Pull Request resolved: https://github.com/facebookincubator/gloo/pull/232

Closes https://github.com/facebookincubator/gloo/issues/227.
Adds 2 methods: abortWaitSend, abortWaitRecv so that we can interrupt a blocking waitRecv or waitSend, per pietern 's suggestion. This is useful if a gloo process is waiting for a message in one thread, but another thread decides to shutdown and wants to stop it gracefully.
The methods set a boolean variable and notify a waiting condition variable. The condition variable predicate is changed to check this condition. The implementation is both for the tcp and uv transport methods. Unit tests are also added.

Differential Revision: D18321454

fbshipit-source-id: 2b0b5af02cf266fd49a4b78d722ff7f2be47bd15