Fix lock inversion issue (#211)
Summary:
Pull Request resolved: https://github.com/facebookincubator/gloo/pull/211

This path is executed when an error happens waiting for a send or recv
operation against an unbound buffer to complete. The unbound buffer
code then signals the context that an error has happened, which in
turn signals all the pair objects that an error has happened.

This is done such that all future operations will fail as well (as we
cannot guarantee the connected context is still usable after an I/O
error or a timeout).

The problem was rooted in the context lock being acquired before each
of the pair locks were, whereas during nominal operation, the pair
lock is acquired before the context lock.

The solution is to remove the context lock entirely, because the
`pairs_` vector is logically const anyway. It is constructed once and
destructed at context destruction. In between, we do not need
exclusion to access it.

Test Plan: Imported from OSS

Differential Revision: D17258597

Pulled By: pietern

fbshipit-source-id: 5b272d770b46a7677f53245bea59521f1ce35dad