Update and move remote pending operation accounting (#226)
Summary:
Accounting for remote pending operations was the responsibility of the
transport specific context class. This is essentially identical across
transports, though, so this commit moves the logic out of the two
transport implementations and into the base class.

Second, the original approach could lead to starvation on
recv-from-any calls. Because pending operations were counted keyed off
the remote rank, low ranks would take precedence over high ranks. This
commit changes that such that ranks are serviced in the same order in
which their send notifications are received.
Pull Request resolved: https://github.com/facebookincubator/gloo/pull/226

Reviewed By: d4l3k

Differential Revision: D18172116

Pulled By: pietern

fbshipit-source-id: b2513478527996594d4a80f4be4e2ee4711fcbc0