Make gloo::transport::tcp::Pair::changeState noexcept(true)
Summary:
If `Pair::listen` encounters an error it calls signal and throw
exception. Signaling an exception means closing the pair, which in
turn calls `Pair::changeState` to move to the CLOSED state. The code
in this function would throw if the current state is equal to
INITIALIZING, which would mask the exception that `Pair::listen`
intended to throw. This is solved by making `Pair::changeState`
noexcept(true) and updating the state selection code in its body.

Differential Revision: D18807092

fbshipit-source-id: a1731bfabbf6c3b1feccdd402a7b5e1b46194dc7