Make gloo::transport::tcp::Pair::changeState noexcept(true) (#235)
Summary:
Pull Request resolved: https://github.com/facebookincubator/gloo/pull/235

If `Pair::listen` encounters an error it calls signal and throw
exception. Signaling an exception means closing the pair, which in
turn calls `Pair::changeState` to move to the CLOSED state. The code
in this function would throw if the current state is equal to
INITIALIZING, which would mask the exception that `Pair::listen`
intended to throw. This is solved by making `Pair::changeState`
noexcept(true) and updating the state selection code in its body.

Reviewed By: jiajunshen

Differential Revision: D18807092

fbshipit-source-id: 2a4c391e82b68362cd0b417d913f175ddc077906