Update on "Use a single listening socket per device"
If Gloo is used with a large number of ranks, the number of listening
sockets grows quadratic with the context size. If you happen to
colocate many ranks on a single machine, you may eventually run out of
ports for listening sockets.

We fix this by using a single listening socket per device instance
instead of using one per pair. Connections to all pair instances are
multiplexed on a single listening socket by adding a sequence number
to the address struct. Upon construction of a new pair, it asks the
listener to return an address with a unique sequence number. When
these addresses are exchanged, both sides of the pair decide which
side is going to be the connection initiator and which one is going to
be the listener. The initiator will connect to the listed address and
write a few bytes containing the sequence number. The listener waits
for a connection to the shared listening socket where it can read that
same sequence number.

This technique is identical to the one used in the uv transport.

Differential Revision: [D18909099](https://our.internmc.facebook.com/intern/diff/D18909099)