Create and use epoll(2) handler base class (#236)
Summary:
Pull Request resolved: https://github.com/facebookincubator/gloo/pull/236

Dispatch to multiple types is needed because we must deal with a
single listening socket on the device instance and I/O for all pair
instances. Before this approach, we'd exclusively deal with `Pair`
instances and didn't need to dispatch events to different types.

Differential Revision: D18909051

Test Plan: Imported from OSS

Pulled By: pietern

fbshipit-source-id: 9222fba1f31deee0b57c22b68a8caa2440d64663