Pass local rank to Pair object (#273)
Summary:
Pull Request resolved: https://github.com/facebookincubator/gloo/pull/273

Currently, the local rank is being calculated but not stored anywhere. The purpose of this diff is to pass and store this local rank in the Pair object. Introduced a protected field in the transport/pair class to store the local rank. Implemented getters and setters for that local rank. Set the local rank after it has been computed for each Pair that is created in connectFullMesh.

Reviewed By: osalpekar

Differential Revision: D25917169

fbshipit-source-id: d932004a445980de75c763b9a893c4a452b8d864