Add benchmarking support for allgather, allgatherv, allgather_ring, and reduce_scatter (#277)
Summary:
Pull Request resolved: https://github.com/facebookincubator/gloo/pull/277

**This Diff:**

Adds benchmarking support for allgather, allgatherv, allgather_ring, and reduce_scatter.

- Added new cases in `RUN_BENCHMARK(T)`
- Updated usages in for `./benchmark --help`
- `allgather_ring` and `reduce_scatter` were implemented already so just added the usage line in `--help` for those
- Renamed `AllgatherBenchmark` to `AllgatherRingBenchmark`
- Created a new classes `AllgatherBenchmark` and `AllgathervBenchmark` which inherit from `Benchmark`. It contains additional fields required to configure the options struct required for the collective.

In `initialize`, we create the required variables and then configure the options struct using those. In `run` we can just call the collective function on these options.

**This Stack:**

The purpose of this stack is to have a comprehensive benchmark for all collectives currently implemented (See list of missing benchmarks in task).

The default `Benchmark::run` function calls the `Algorithm::run` function associated with the `algorithm_` we set in initialize. The remaining missing collectives (with the exception of `AllreduceLocal`) do not inherit from `Algorithm` class which means that we need to override both the `initialize` and `run` functions.

Reviewed By: osalpekar

Differential Revision: D26020178

fbshipit-source-id: ccf9b16f3cf4d7e70722ae7c488b9addd49a736f