Add benchmarking support for broadcast, reduce, and scatter (#278)
Summary:
Pull Request resolved: https://github.com/facebookincubator/gloo/pull/278

**This Diff:**

Adds benchmarking support for broadcast, reduce, and scatter.

- Added new cases in `RUN_BENCHMARK(T)`
- Updated usages in for `./benchmark --help`
- Created a new classes `BroadcastBenchmark`, `ReduceBenchmark` and `ScatterBenchmark` which inherit from `Benchmark`. It contains additional fields required to configure the options struct required for the collective.

In `initialize`, we create the required variables and then configure the corresponding options struct using those. In `run` we can just call the collective function on these options.

**This Stack:**

The purpose of this stack is to have a comprehensive benchmark for all collectives currently implemented (See list of missing benchmarks in task).

The default `Benchmark::run` function calls the `Algorithm::run` function associated with the `algorithm_` we set in initialize. The remaining missing collectives (with the exception of `AllreduceLocal`) do not inherit from `Algorithm` class which means that we need to override both the `initialize` and `run` functions.

Reviewed By: osalpekar

Differential Revision: D26020282

fbshipit-source-id: a87aabe918e15f5085da76a069bb5c0b5a020f7a