Add benchmarking support for alltoallv (#276)
Summary:
Pull Request resolved: https://github.com/facebookincubator/gloo/pull/276

**This Diff:**

Adds benchmarking support for alltoallv.

- Added new cases in `RUN_BENCHMARK(T)`
- Updated usages in for `./benchmark --help`
- Support for `alltoallv`:

Created a new class `AllToAllvBenchmark` which inherits from `Benchmark`. It contains additional fields required to configure the `AlltoallvOptions`.

In `initialize`, we create vectors the same way done in `alltoallv_test.cc` and then configure the options struct using those vectors. In `run` we can just call the collective function on these options.

**This Stack:**

The purpose of this stack is to have a comprehensive benchmark for all collectives currently implemented (See list of missing benchmarks in task).

The default `Benchmark::run` function calls the `Algorithm::run` function associated with the `algorithm_` we set in initialize. The remaining missing collectives (with the exception of `AllreduceLocal`) do not inherit from `Algorithm` class which means that we need to override both the `initialize` and `run` functions.

Reviewed By: osalpekar

Differential Revision: D25987688

fbshipit-source-id: 0333f02deeb65bc5c43b0759b9b09250f62ef71f