Add benchmarking support for p2p communications (alternating msg size)
Summary:
**This diff:**
The purpose of this diff is to add benchmarks for send/recv and isend/irecv where it sends many messages of alternating size (large and small)

Since the code is very similar for this benchmark and the existing benchmark SendRecvStressBenchmark, I modified the existing class to have the extra option to alternate message sizes when sending (rather than creating a whole new class).

Also updated the `--help` to reflect these new benchmarks.

**This stack:**
The purpose of this stack is to add support for various benchmarks for p2p communications (send/recv/isend/irecv)

Differential Revision: D26316098

fbshipit-source-id: 2b9a73e6b422c3c64313c5ed3fc7e38530387ebb