Add support for minimum run time when using benchmark (#296)
Summary:
Pull Request resolved: https://github.com/facebookincubator/gloo/pull/296

The purpose of this diff is to add an option for minimum run time when running the benchmark. Currently, when iteration count is not specified, the benchmark takes the `iterationTimeNanos` and divides it by the median time spent during warmup runs to get an estimate of the number of iterations. The user is able to specify the `iterationTimeNanos` but the benchmark does not actually run for that long.

Instead, we want the user to be able to specify the minimum time that the benchmark runs for.

Referenced Google's Benchmark: https://github.com/google/benchmark/blob/5c43112eb3d70cec10784d8d1f4373bc7397836b/src/benchmark_runner.cc#L252

Changes:
- Add new loop logic so that benchmark will re-run with increased iteration count if total time spent is less than the specified amount
- Add new class function `Samples::sum` so we can get total sum without having to sort first using `Distribution` class
- Add warmup iterations as an option for the user to specify
- Rename iterationTimeNanos to minIterationTimeNanos to be more clear
- Clarify the usage of certain options in `--help`
- Add max number of iterations (constant) when running for a specified amount of time

Reviewed By: mingzhe09088

Differential Revision: D26644215

fbshipit-source-id: 2cc5c155249786a53df94664bde95f3e94d9aa93